itertest: add internal iterator datadriven helper
Move the datadriven internal iterator helper out of the pebble package's
data_test.go file and into a new internal/itertest package so that it may be
used to drive unit tests outside of the primary pebble package.
