+++ data_test.go
@@ -34,21 +34,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-type iterCmdOpts struct {
-	verboseKey bool
-	stats      *base.InternalIteratorStats
-}
-
-type iterCmdOpt func(*iterCmdOpts)
-
-func iterCmdVerboseKey(opts *iterCmdOpts) { opts.verboseKey = true }
-
-func iterCmdStats(stats *base.InternalIteratorStats) iterCmdOpt {
-	return func(opts *iterCmdOpts) {
-		opts.stats = stats
-	}
-}
-
 func runGetCmd(t testing.TB, td *datadriven.TestData, d *DB) string {
 	snap := Snapshot{
 		db:     d,
@@ -392,122 +377,6 @@ func writeRangeKeys(b io.Writer, iter *Iterator) {
 		}
 		fmt.Fprintf(b, " %s=%s", rangeKeys[j].Suffix, rangeKeys[j].Value)
 	}
-}
-
-func runInternalIterCmd(
-	t *testing.T, d *datadriven.TestData, iter internalIterator, opts ...iterCmdOpt,
-) string {
-	var o iterCmdOpts
-	for _, opt := range opts {
-		opt(&o)
-	}
-
-	getKV := func(key *InternalKey, val LazyValue) (*InternalKey, []byte) {
-		v, _, err := val.Value(nil)
-		require.NoError(t, err)
-		return key, v
-	}
-	var b bytes.Buffer
-	var prefix []byte
-	for _, line := range strings.Split(d.Input, "\n") {
-		parts := strings.Fields(line)
-		if len(parts) == 0 {
-			continue
-		}
-		var key *InternalKey
-		var value []byte
-		switch parts[0] {
-		case "seek-ge":
-			if len(parts) < 2 || len(parts) > 3 {
-				return "seek-ge <key> [<try-seek-using-next>]\n"
-			}
-			prefix = nil
-			var flags base.SeekGEFlags
-			if len(parts) == 3 {
-				if trySeekUsingNext, err := strconv.ParseBool(parts[2]); err != nil {
-					return err.Error()
-				} else if trySeekUsingNext {
-					flags = flags.EnableTrySeekUsingNext()
-				}
-			}
-			key, value = getKV(iter.SeekGE([]byte(strings.TrimSpace(parts[1])), flags))
-		case "seek-prefix-ge":
-			if len(parts) != 2 && len(parts) != 3 {
-				return "seek-prefix-ge <key> [<try-seek-using-next>]\n"
-			}
-			prefix = []byte(strings.TrimSpace(parts[1]))
-			var flags base.SeekGEFlags
-			if len(parts) == 3 {
-				if trySeekUsingNext, err := strconv.ParseBool(parts[2]); err != nil {
-					return err.Error()
-				} else if trySeekUsingNext {
-					flags = flags.EnableTrySeekUsingNext()
-				}
-			}
-			key, value = getKV(iter.SeekPrefixGE(prefix, prefix /* key */, flags))
-		case "seek-lt":
-			if len(parts) != 2 {
-				return "seek-lt <key>\n"
-			}
-			prefix = nil
-			key, value = getKV(iter.SeekLT([]byte(strings.TrimSpace(parts[1])), base.SeekLTFlagsNone))
-		case "first":
-			prefix = nil
-			key, value = getKV(iter.First())
-		case "last":
-			prefix = nil
-			key, value = getKV(iter.Last())
-		case "next":
-			key, value = getKV(iter.Next())
-		case "prev":
-			key, value = getKV(iter.Prev())
-		case "set-bounds":
-			if len(parts) <= 1 || len(parts) > 3 {
-				return "set-bounds lower=<lower> upper=<upper>\n"
-			}
-			var lower []byte
-			var upper []byte
-			for _, part := range parts[1:] {
-				arg := strings.Split(strings.TrimSpace(part), "=")
-				switch arg[0] {
-				case "lower":
-					lower = []byte(arg[1])
-				case "upper":
-					upper = []byte(arg[1])
-				default:
-					return fmt.Sprintf("set-bounds: unknown arg: %s", arg)
-				}
-			}
-			iter.SetBounds(lower, upper)
-			continue
-		case "stats":
-			if o.stats != nil {
-				// The timing is non-deterministic, so set to 0.
-				o.stats.BlockReadDuration = 0
-				fmt.Fprintf(&b, "%+v\n", *o.stats)
-			}
-			continue
-		case "reset-stats":
-			if o.stats != nil {
-				*o.stats = base.InternalIteratorStats{}
-			}
-			continue
-		default:
-			return fmt.Sprintf("unknown op: %s", parts[0])
-		}
-		if key != nil {
-			if o.verboseKey {
-				fmt.Fprintf(&b, "%s:%s\n", key, value)
-			} else {
-				fmt.Fprintf(&b, "%s:%s\n", key.UserKey, value)
-			}
-		} else if err := iter.Error(); err != nil {
-			fmt.Fprintf(&b, "err=%v\n", err)
-		} else {
-			fmt.Fprintf(&b, ".\n")
-		}
-	}
-	return b.String()
 }
 
 func runBatchDefineCmd(d *datadriven.TestData, b *Batch) error {
